+++ lib/basedriver/capabilities.js
@@ -2,7 +2,7 @@ import _ from 'lodash';
 import { validator } from './desired-caps';
 import { util } from 'appium-support';
 import log from './logger';
-import { errors } from '../mjsonwp/errors';
+import { errors } from '../protocol/errors';
 
 // Takes primary caps object and merges it into a secondary caps object.
 // (see https://www.w3.org/TR/webdriver/#dfn-merging-capabilities)
