+++ lib/basedriver/commands/session.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 import log from '../logger';
-import { errors } from '../../mjsonwp';
+import { errors } from '../../protocol';
 import UUID from 'uuid-js';
 import { processCapabilities } from '../capabilities';
