+++ lib/basedriver/commands/timeout.js
@@ -3,7 +3,7 @@ import { waitForCondition } from 'asyncbox';
 import B from 'bluebird';
 import _ from 'lodash';
 import { util } from 'appium-support';
-import { errors } from '../../mjsonwp';
+import { errors } from '../../protocol';
 import BaseDriver from "../driver";
