+++ lib/basedriver/driver.js
@@ -1,4 +1,4 @@
-import { MobileJsonWireProtocol, errors } from '../mjsonwp';
+import { Protocol, errors } from '../protocol';
 import os from 'os';
 import commands from './commands';
 import helpers from './helpers';
@@ -18,7 +18,7 @@ const EVENT_SESSION_START = 'newSessionStarted';
 const EVENT_SESSION_QUIT_START = 'quitSessionRequested';
 const EVENT_SESSION_QUIT_DONE = 'quitSessionFinished';
 
-class BaseDriver extends MobileJsonWireProtocol {
+class BaseDriver extends Protocol {
 
   constructor (opts = {}, shouldValidateCaps = true) {
     super();
