+++ lib/express/crash.js
@@ -1,4 +1,4 @@
-import { errors } from '../mjsonwp';
+import { errors } from '../protocol';
 
 
 function produceError () {
