+++ lib/express/middleware.js
@@ -1,5 +1,5 @@
 import log from './logger';
-import { errors } from '../mjsonwp';
+import { errors } from '../protocol';
 
 
 function allowCrossDomain (req, res, next) {
