+++ lib/jsonwp-proxy/proxy.js
@@ -2,7 +2,7 @@ import _ from 'lodash';
 import { logger, util } from 'appium-support';
 import request from 'request-promise';
 import { getSummaryByCode } from '../jsonwp-status/status';
-import { errors } from '../mjsonwp/errors';
+import { errors } from '../protocol/errors';
 
 
 const log = logger.getLogger('JSONWP Proxy');
