+++ lib/protocol/errors.js
