+++ lib/protocol/index.js
@@ -1,11 +1,11 @@
 // transpile:main
 
-import { MJSONWP as MobileJsonWireProtocol, isSessionCommand,
-         routeConfiguringFunction } from './mjsonwp';
+import { Protocol, isSessionCommand,
+         routeConfiguringFunction } from './protocol';
 import { NO_SESSION_ID_COMMANDS, ALL_COMMANDS, METHOD_MAP,
          routeToCommandName } from './routes';
 import { errors, isErrorType, errorFromCode } from './errors';
 
-export { MobileJsonWireProtocol, routeConfiguringFunction, errors, isErrorType,
+export { Protocol, routeConfiguringFunction, errors, isErrorType,
          errorFromCode, ALL_COMMANDS, METHOD_MAP, routeToCommandName,
          NO_SESSION_ID_COMMANDS, isSessionCommand };
