+++ lib/protocol/protocol.js
@@ -7,7 +7,9 @@ import { renameKey } from '../basedriver/helpers';
 import B from 'bluebird';
 
 
-const log = logger.getLogger('MJSONWP');
+const mjsonwpLog = logger.getLogger('MJSONWP');
+const w3cLog = logger.getLogger('W3C');
+
 const JSONWP_SUCCESS_STATUS_CODE = 0;
 // TODO: Make this value configurable as a server side capability
 const LOG_OBJ_LENGTH = 1024; // MAX LENGTH Logged to file / console
@@ -15,7 +17,11 @@ const LOG_OBJ_LENGTH = 1024; // MAX LENGTH Logged to file / console
 const MJSONWP_ELEMENT_KEY = 'ELEMENT';
 const W3C_ELEMENT_KEY = 'element-6066-11e4-a52e-4f735466cecf';
 
-class MJSONWP {}
+class Protocol {}
+
+function getLogByProtocol (driver) {
+  return driver.isW3CProtocol() ? w3cLog : mjsonwpLog;
+}
 
 function isSessionCommand (command) {
   return !_.includes(NO_SESSION_ID_COMMANDS, command);
@@ -192,6 +198,10 @@ function buildHandler (app, method, path, spec, driver, isSessCmd) {
     let httpResBody = {};
     let httpStatus = 200;
     let newSessionId;
+
+    // get the appropriate logger depending on the protocol that is being used
+    const log = getLogByProtocol(driver);
+
     try {
       // if this is a session command but we don't have a session,
       // error out early (especially before proxying)
@@ -238,6 +248,7 @@ function buildHandler (app, method, path, spec, driver, isSessCmd) {
       if (validators[spec.command]) {
         validators[spec.command](...args);
       }
+
       // run the driver command wrapped inside the argument validators
       log.debug(`Calling ${driver.constructor.name}.${spec.command}() with args: ` +
                 _.truncate(JSON.stringify(args), {length: LOG_OBJ_LENGTH}));
@@ -378,7 +389,7 @@ function driverShouldDoJwpProxy (driver, req, command) {
 }
 
 async function doJwpProxy (driver, req, res) {
-  log.info('Driver proxy active, passing request on via HTTP proxy');
+  getLogByProtocol(driver).info('Driver proxy active, passing request on via HTTP proxy');
 
   // check that the inner driver has a proxy function
   if (!driver.canProxy(req.params.sessionId)) {
@@ -396,4 +407,4 @@ async function doJwpProxy (driver, req, res) {
 }
 
 
-export { MJSONWP, routeConfiguringFunction, isSessionCommand, MJSONWP_ELEMENT_KEY, W3C_ELEMENT_KEY };
+export { Protocol, routeConfiguringFunction, isSessionCommand, MJSONWP_ELEMENT_KEY, W3C_ELEMENT_KEY };
